add header files to major modules
