+++ libgadget/Makefile
@@ -8,11 +8,12 @@ include $(CONFIG)
 
 INCL = config.h \
 densitykernel.h  endrun.h    forcetree.h  interp.h event.h hci.h \
-mymalloc.h    paramset.h  petapm.h         proto.h    timebinmgr.h  treewalk.h \
+mymalloc.h    paramset.h  petapm.h         run.h  timebinmgr.h  treewalk.h \
 allvars.h  partmanager.h  cooling.h         domain.h         exchange.h  slotsmanager.h     \
+checkpoint.h \
  openmpsort.h  peano.h     physconst.h      sfr_eff.h  timefac.h     utils-string.h \
 blackhole.h  cosmology.h       drift.h          fof.h       gravshort.h  memory.h   \
-param.h       petaio.h    powerspectrum.h  system.h   timestep.h    walltime.h \
+ petaio.h    powerspectrum.h  system.h   timestep.h    walltime.h \
 kspace-neutrinos/omega_nu_single.h kspace-neutrinos/delta_tot_table.h
 
 include ../Makefile.rules
@@ -31,18 +32,18 @@ include ../Makefile.version
 GADGET_OBJS =  \
 	 gdbtools.o hci.o\
 	 fof.o fofpetaio.o petaio.o \
-	 param.o paramset.o \
+	 paramset.o \
 	 domain.o exchange.o slotsmanager.o allvars.o partmanager.o \
 	 blackhole.o timebinmgr.o \
-	 run.o drift.o begrun.o global.o \
-	 timestep.o init.o io.o sfr_eff.o \
+	 run.o drift.o stats.o \
+	 timestep.o init.o checkpoint.o sfr_eff.o \
 	 cooling.o \
 	 density.o \
 	 treewalk.o cosmology.o \
 	 gravshort-tree.o gravshort-pair.o hydra.o  timefac.o \
 	 gravpm.o powerspectrum.o \
 	 forcetree.o peano.o \
-	 petapm.o longrange.o \
+	 petapm.o gravity.o \
 	 densitykernel.o interp.o lightcone.o walltime.o\
 	 runtests.o \
      kspace-neutrinos/omega_nu_single.o \
