+++ libgadget/checkpoint.h
@@ -1,0 +1,4 @@
+void write_checkpoint(int write_snapshot, int write_fof);
+void dump_snapshot(void);
+int find_last_snapnum();
+
