+++ libgadget/density.h
@@ -1,0 +1,2 @@
+void density(void);
+void density_update(void);
