+++ libgadget/lightcone.c
@@ -10,7 +10,6 @@
 #include <sys/types.h>
 #include "allvars.h"
 #include "partmanager.h"
-#include "proto.h"
 #include "cosmology.h"
 #include "endrun.h"
 #include "system.h"
