+++ libgadget/run.c
@@ -7,11 +7,16 @@
 #include <ctype.h>
 
 #include "allvars.h"
-#include "proto.h"
+#include "gravity.h"
+#include "density.h"
 #include "domain.h"
+#include "run.h"
 #include "cooling.h"
 #include "mymalloc.h"
 #include "endrun.h"
+#include "checkpoint.h"
+#include "petaio.h"
+#include "utils-string.h"
 #include "timestep.h"
 #include "system.h"
 #include "drift.h"
@@ -21,6 +26,11 @@
 #include "sfr_eff.h"
 #include "slotsmanager.h"
 #include "hci.h"
+
+#include "kspace-neutrinos/delta_tot_table.h"
+
+void energy_statistics(void); /* stats.c only used here */
+void init(int snapnum); /* init.c only used here */
 
 /*! \file run.c
  *  \brief  iterates over timesteps, main loop
@@ -33,6 +43,79 @@
  */
 static void compute_accelerations(int is_PM, int FirstStep, int GasEnabled);
 static void write_cpu_log(int NumCurrentTiStep);
+
+/*! \file begrun.c
+ *  \brief initial set-up of a simulation run
+ *
+ *  This file contains various functions to initialize a simulation run. In
+ *  particular, the parameterfile is read in and parsed, the initial
+ *  conditions or restart files are read, and global variables are initialized
+ *  to their proper values.
+ */
+
+
+static void set_units();
+static void set_softenings();
+
+static void
+open_outputfiles(int RestartSnapNum);
+
+static void
+close_outputfiles(void);
+
+/*Defined in gravpm.c*/
+extern _delta_tot_table delta_tot_table;
+extern _transfer_init_table transfer_init;
+
+/*! This function performs the initial set-up of the simulation. First, the
+ *  parameterfile is set, then routines for setting units, reading
+ *  ICs/restart-files are called, auxialiary memory is allocated, etc.
+ */
+void begrun(int RestartSnapNum)
+{
+
+    hci_init(HCI_DEFAULT_MANAGER, All.OutputDir, All.TimeLimitCPU, All.AutoSnapshotTime);
+
+    petaio_init();
+    walltime_init(&All.CT);
+
+    petaio_read_header(RestartSnapNum);
+
+    slots_init();
+    /* Enable the slots: stars and BHs are allocated if there are some,
+     * or if some will form*/
+    if(All.NTotalInit[0] > 0)
+        slots_set_enabled(0, sizeof(struct sph_particle_data));
+    if(All.StarformationOn || All.NTotalInit[4] > 0)
+        slots_set_enabled(4, sizeof(struct star_particle_data));
+    if(All.BlackHoleOn || All.NTotalInit[5] > 0)
+        slots_set_enabled(5, sizeof(struct bh_particle_data));
+
+    set_softenings();
+    set_units();
+
+#ifdef DEBUG
+    write_pid_file(All.OutputDir);
+    enable_core_dumps_and_fpu_exceptions();
+#endif
+    InitCool();
+
+#if defined(SFR)
+    init_clouds();
+#endif
+
+    grav_init();
+
+    set_random_numbers(All.RandomSeed);
+
+    init(RestartSnapNum);			/* ... read in initial model */
+
+#ifdef LIGHTCONE
+    lightcone_init(All.Time);
+#endif
+
+    open_outputfiles(RestartSnapNum);
+}
 
 void run(void)
 {
@@ -174,6 +257,8 @@ void run(void)
             apply_PM_half_kick();
         }
     }
+
+    close_outputfiles();
 }
 
 /*! This routine computes the accelerations for all active particles.  First, the gravitational forces are
@@ -280,3 +365,192 @@ void write_cpu_log(int NumCurrentTiStep)
         fflush(FdCPU);
     }
 }
+
+/*!  This function opens various log-files that report on the status and
+ *   performance of the simulstion. On restart from restart-files
+ *   (start-option 1), the code will append to these files.
+ */
+static void
+open_outputfiles(int RestartSnapNum)
+{
+    const char mode[3]="a+";
+    char * buf;
+    char * postfix;
+
+    if(ThisTask != 0) {
+        /* only the root processors writes to the log files */
+        return;
+    }
+
+    if(RestartSnapNum != -1) {
+        postfix = fastpm_strdup_printf("-R%03d", RestartSnapNum);
+    } else {
+        postfix = fastpm_strdup_printf("%s", "");
+    }
+
+    buf = fastpm_strdup_printf("%s/%s%s", All.OutputDir, All.CpuFile, postfix);
+    fastpm_path_ensure_dirname(buf);
+    if(!(FdCPU = fopen(buf, mode)))
+        endrun(1, "error in opening file '%s'\n", buf);
+    free(buf);
+
+    if(All.OutputEnergyDebug) {
+        buf = fastpm_strdup_printf("%s/%s%s", All.OutputDir, All.EnergyFile, postfix);
+        fastpm_path_ensure_dirname(buf);
+        if(!(FdEnergy = fopen(buf, mode)))
+            endrun(1, "error in opening file '%s'\n", buf);
+        free(buf);
+    }
+
+#ifdef SFR
+    buf = fastpm_strdup_printf("%s/%s%s", All.OutputDir, "sfr.txt", postfix);
+    fastpm_path_ensure_dirname(buf);
+    if(!(FdSfr = fopen(buf, mode)))
+        endrun(1, "error in opening file '%s'\n", buf);
+    free(buf);
+#endif
+
+#ifdef BLACK_HOLES
+    buf = fastpm_strdup_printf("%s/%s%s", All.OutputDir, "blackholes.txt", postfix);
+    fastpm_path_ensure_dirname(buf);
+    if(!(FdBlackHoles = fopen(buf, mode)))
+        endrun(1, "error in opening file '%s'\n", buf);
+    free(buf);
+#endif
+
+}
+
+
+/*!  This function closes the global log-files.
+*/
+static void
+close_outputfiles(void)
+{
+
+    if(ThisTask != 0)		/* only the root processors writes to the log files */
+        return;
+
+    fclose(FdCPU);
+    if(All.OutputEnergyDebug)
+        fclose(FdEnergy);
+
+#ifdef SFR
+    fclose(FdSfr);
+#endif
+
+#ifdef BLACK_HOLES
+    fclose(FdBlackHoles);
+#endif
+}
+
+/*! Computes conversion factors between internal code units and the
+ *  cgs-system.
+ */
+static void
+set_units(void)
+{
+    double meanweight;
+
+    All.UnitTime_in_s = All.UnitLength_in_cm / All.UnitVelocity_in_cm_per_s;
+    All.UnitTime_in_Megayears = All.UnitTime_in_s / SEC_PER_MEGAYEAR;
+
+    All.G = GRAVITY / pow(All.UnitLength_in_cm, 3) * All.UnitMass_in_g * pow(All.UnitTime_in_s, 2);
+
+    All.UnitDensity_in_cgs = All.UnitMass_in_g / pow(All.UnitLength_in_cm, 3);
+    All.UnitPressure_in_cgs = All.UnitMass_in_g / All.UnitLength_in_cm / pow(All.UnitTime_in_s, 2);
+    All.UnitCoolingRate_in_cgs = All.UnitPressure_in_cgs / All.UnitTime_in_s;
+    All.UnitEnergy_in_cgs = All.UnitMass_in_g * pow(All.UnitLength_in_cm, 2) / pow(All.UnitTime_in_s, 2);
+
+    /* convert some physical input parameters to internal units */
+
+    All.CP.Hubble = HUBBLE * All.UnitTime_in_s;
+    init_cosmology(&All.CP, All.TimeInit, All.BoxSize * All.UnitLength_in_cm);
+    /*Initialise the hybrid neutrinos, after Omega_nu*/
+    if(All.HybridNeutrinosOn)
+        init_hybrid_nu(&All.CP.ONu.hybnu, All.CP.MNu, All.HybridVcrit, C/1e5, All.HybridNuPartTime, All.CP.ONu.kBtnu);
+
+    meanweight = 4.0 / (1 + 3 * HYDROGEN_MASSFRAC);	/* note: assuming NEUTRAL GAS */
+
+    All.MinEgySpec = 1 / meanweight * (1.0 / GAMMA_MINUS1) * (BOLTZMANN / PROTONMASS) * All.MinGasTemp;
+    All.MinEgySpec *= All.UnitMass_in_g / All.UnitEnergy_in_cgs;
+
+#ifdef SFR
+
+    All.OverDensThresh =
+        All.CritOverDensity * All.CP.OmegaBaryon * 3 * All.CP.Hubble * All.CP.Hubble / (8 * M_PI * All.G);
+
+    All.PhysDensThresh = All.CritPhysDensity * PROTONMASS / HYDROGEN_MASSFRAC / All.UnitDensity_in_cgs;
+
+    All.EgySpecCold = 1 / meanweight * (1.0 / GAMMA_MINUS1) * (BOLTZMANN / PROTONMASS) * All.TempClouds;
+    All.EgySpecCold *= All.UnitMass_in_g / All.UnitEnergy_in_cgs;
+
+    meanweight = 4 / (8 - 5 * (1 - HYDROGEN_MASSFRAC));	/* note: assuming FULL ionization */
+
+    All.EgySpecSN = 1 / meanweight * (1.0 / GAMMA_MINUS1) * (BOLTZMANN / PROTONMASS) * All.TempSupernova;
+    All.EgySpecSN *= All.UnitMass_in_g / All.UnitEnergy_in_cgs;
+
+    if(All.WindOn) {
+        if(HAS(All.WindModel, WIND_FIXED_EFFICIENCY)) {
+            All.WindSpeed = sqrt(2 * All.WindEnergyFraction * All.FactorSN * All.EgySpecSN / (1 - All.FactorSN) / All.WindEfficiency);
+            message(0, "Windspeed: %g\n", All.WindSpeed);
+        } else {
+            All.WindSpeed = sqrt(2 * All.WindEnergyFraction * All.FactorSN * All.EgySpecSN / (1 - All.FactorSN) / 1.0);
+            message(0, "Reference Windspeed: %g\n", All.WindSigma0 * All.WindSpeedFactor);
+        }
+    }
+#endif
+
+    message(0, "Hubble (internal units) = %g\n", All.CP.Hubble);
+    message(0, "G (internal units) = %g\n", All.G);
+    message(0, "UnitLengh_in_cm = %g \n", All.UnitLength_in_cm);
+    message(0, "UnitMass_in_g = %g \n", All.UnitMass_in_g);
+    message(0, "UnitTime_in_s = %g \n", All.UnitTime_in_s);
+    message(0, "UnitVelocity_in_cm_per_s = %g \n", All.UnitVelocity_in_cm_per_s);
+    message(0, "UnitDensity_in_cgs = %g \n", All.UnitDensity_in_cgs);
+    message(0, "UnitEnergy_in_cgs = %g \n", All.UnitEnergy_in_cgs);
+    message(0, "Photon density OmegaG = %g\n",All.CP.OmegaG);
+    if(!All.MassiveNuLinRespOn)
+        message(0, "Massless Neutrino density OmegaNu0 = %g\n",get_omega_nu(&All.CP.ONu, 1));
+    message(0, "Curvature density OmegaK = %g\n",All.CP.OmegaK);
+    if(All.CP.RadiationOn) {
+        /* note that this value is inaccurate if there is massive neutrino. */
+        double OmegaTot = All.CP.OmegaG + All.CP.OmegaK + All.CP.Omega0 + All.CP.OmegaLambda;
+        if(!All.MassiveNuLinRespOn)
+            OmegaTot += get_omega_nu(&All.CP.ONu, 1);
+        message(0, "Radiation is enabled in Hubble(a). "
+               "Following CAMB convention: Omega_Tot - 1 = %g\n", OmegaTot - 1);
+    }
+    message(0, "\n");
+}
+
+/*! This function sets the (comoving) softening length of all particle
+ *  types in the table All.SofteningTable[...].  We check that the physical
+ *  softening length is bounded by the Softening-MaxPhys values.
+ */
+static void
+set_softenings()
+{
+    int i;
+
+    for(i = 0; i < 6; i ++)
+        GravitySofteningTable[i] = All.GravitySoftening * All.MeanSeparation[1];
+
+    /* 0: Gas is collisional */
+    GravitySofteningTable[0] = All.GravitySofteningGas * All.MeanSeparation[1];
+
+    All.MinGasHsml = All.MinGasHsmlFractional * GravitySofteningTable[1];
+
+    for(i = 0; i < 6; i ++) {
+        message(0, "GravitySoftening[%d] = %g\n", i, GravitySofteningTable[i]);
+    }
+
+    double minsoft = 0;
+    for(i = 0; i<6; i++) {
+        if(GravitySofteningTable[i] <= 0) continue;
+        if(minsoft == 0 || minsoft > GravitySofteningTable[i])
+            minsoft = GravitySofteningTable[i];
+    }
+    /* FIXME: make this a parameter. */
+    All.TreeNodeMinSize = 1.0e-3 * 2.8 * minsoft;
+}
+
