+++ libgadget/run.h
@@ -1,0 +1,9 @@
+#ifndef RUN_H
+#define RUN_H
+
+void begrun(int RestartSnapNum);
+
+void run(void);
+void runtests(void);
+
+#endif
