+++ libgadget/runtests.c
@@ -9,8 +9,8 @@
 #include "allvars.h"
 #include "partmanager.h"
 #include "forcetree.h"
-#include "proto.h"
 #include "cooling.h"
+#include "gravity.h"
 #include "mymalloc.h"
 #include "endrun.h"
 #include "petaio.h"
