+++ libgadget/stats.c
