+++ libgadget/timestep.c
@@ -5,13 +5,12 @@
 #include <math.h>
 #include "allvars.h"
 #include "domain.h"
-#include "openmpsort.h"
-#include "proto.h"
 #include "timefac.h"
 #include "cosmology.h"
 #include "cooling.h"
 #include "mymalloc.h"
 #include "endrun.h"
+#include "checkpoint.h"
 #include "slotsmanager.h"
 #include "partmanager.h"
 #include "system.h"
